+++ HISTORY.rst
@@ -3,6 +3,8 @@
 - Document stubs without a label will now display their ID as the label.
   This allows documents without files or versions to be accessible via the
   user interface.
+- Return the event type subscription list sorted by namespace label and event
+  type label.
 
 3.5.3 (2020-11-11)
 ==================